pyproject.toml
[project]
name = "claude"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Donald Thompson", email = "witt3rd@witt3rd.com" }]
dependencies = [
    "anthropic>=0.42.0",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/claude"]

[tool.rye.scripts]
api = "uvicorn src.claude.main:app --reload --port 9000"
